render books

+++ app/src/main/java/ro/holdone/swissborg/server/model/BookEntry.kt
@@ -1,6 +1,6 @@
 package ro.holdone.swissborg.server.model
 
-data class BookEntry(val price: Float, val count: Float, val amount: Float) {
+data class BookEntry(val price: Float, val count: Int, val amount: Float) {
     override fun toString(): String {
         return "BookEntry(price=$price, count=$count, amount=$amount)"
     }
